Dev3

+++ src/main/java/cpsc326/Interpreter.java
@@ -19,7 +19,7 @@ class Interpreter implements Expr.Visitor<Object>, Stmt.Visitor<Void>
             public int arity() {return 0;};
 
             @Override
-            public Object call(Interpreter interpreter, List<Object> arguments) {return (System.currentTimeMillis()/1000);};
+            public Object call(Interpreter interpreter, List<Object> arguments) {return (double)(System.currentTimeMillis()/1000);};
         });
     }
